requirements.txt
aiodns==3.0.0
aiohttp==3.7.4.post0
appdirs==1.4.4
apturl==0.5.2
asgiref==3.4.1
astroid==2.6.2
async-timeout==3.0.1
attrs==21.2.0
backcall==0.1.0
backports.entry-points-selectable==1.1.1
bcrypt==3.1.7
beautifulsoup4==4.9.3
bleach==3.1.1
blinker==1.4
Brlapi==0.7.0
bs4==0.0.1
-e git+https://github.com/kivy/buildozer@04f5d0076ea7931fc8aa0588f6a4c81810e2699e#egg=buildozer
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
Cython==0.29.19
dbus-python==1.2.16
dearpygui==1.1.3
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==3.2.5
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.12.4
dnslib==0.9.16
dnspython==2.0.0
docstring-to-markdown==0.10
docutils==0.18.1
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
Faker==13.0.0
fasteners==0.14.1
ffpyplayer==4.3.2
filelock==3.4.2
flake8==4.0.1
fonttools==4.28.2
future==0.18.2
gunicorn==20.0.4
html5lib==1.0.1
httplib2==0.14.0
icmplib==3.0
idna==2.8
importlib-metadata==1.5.0
iniconfig==1.1.1
ipwhois==1.2.0
ipykernel==5.2.0
ipython==7.13.0
ipython_genutils==0.2.0
ipywidgets==6.0.0
isort==5.9.1
jedi==0.18.0
jedi-language-server==0.35.0
Jinja2==2.10.1
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
keyring==18.0.1
Kivy==2.0.0
Kivy-examples==2.0.0
Kivy-Garden==0.1.4
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-object-proxy==1.6.0
lockfile==0.12.2
louis==3.12.0
lxml==4.6.3
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==1.1.0
matplotlib==3.5.0
mccabe==0.6.1
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
multidict==5.1.0
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.4
notebook==6.0.3
numpy==1.21.0
oauthlib==3.1.0
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
pandocfilters==1.4.2
paramiko==2.6.0
parso==0.8.3
pep517==0.6.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.3.1
platformdirs==2.4.1
pluggy==1.0.0
prometheus-client==0.7.1
prompt-toolkit==2.0.10
protobuf==3.6.1
psycopg2==2.9.3
psycopg2-binary==2.9.1
py==1.11.0
pycairo==1.16.2
pycares==4.0.0
pycodestyle==2.8.0
pycparser==2.20
pycups==1.9.73
pydantic==1.9.0
pyflakes==2.4.0
pygame==2.1.2
pygls==0.11.3
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pylint==2.9.3
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyparsing==3.0.6
pyRFC3339==1.1
pyrsistent==0.15.5
PySimpleGUI==4.45.0
pytest==6.2.5
python-apt==2.0.0+ubuntu0.20.4.6
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-for-android==2021.9.5
pytoml==0.1.21
pytz==2021.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
requests==2.22.0
requests-file==1.5.1
requests-unixsocket==0.2.0
scipy==1.7.1
SecretStorage==2.3.1
selenium==3.141.0
Send2Trash==1.5.0
setuptools-scm==6.3.2
sh==1.14.2
simplejson==3.16.0
six==1.14.0
soupsieve==2.2.1
sqlparse==0.4.1
subprocess32==3.5.4
system-service==0.3
systemd-python==234
terminado==0.8.2
testpath==0.4.4
tldextract==3.1.0
toml==0.10.2
tomli==1.2.2
tornado==5.1.1
traitlets==4.3.3
typeguard==2.13.3
typing-extensions==3.10.0.0
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
usb-creator==0.3.7
validate-docbr==1.8.2
virtualenv==20.13.0
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
whitenoise==6.0.0
widgetsnbextension==2.0.0
wrapt==1.12.1
xkit==0.0.0
yarl==1.6.3
zipp==1.0.0
